Support fixed database role db_ddladmin (#3130)
Support fixed database role db_ddladmin

Engine PR : amazon-aurora/postgresql_modified_for_babelfish#101

Task : BABEL-5116

+++ contrib/babelfishpg_tsql/src/schemacmds.c
@@ -76,9 +76,6 @@ del_ns_ext_info(const char *schemaname, bool missing_ok)
 	ScanKeyData scanKey;
 	SysScanDesc scan;
 
-	if (get_namespace_oid(schemaname, missing_ok) == InvalidOid)
-		return;
-
 	rel = table_open(namespace_ext_oid, RowExclusiveLock);
 	ScanKeyInit(&scanKey,
 				Anum_namespace_ext_namespace,
@@ -93,9 +90,10 @@ del_ns_ext_info(const char *schemaname, bool missing_ok)
 	{
 		systable_endscan(scan);
 		table_close(rel, RowExclusiveLock);
-		ereport(ERROR,
-				(errcode(ERRCODE_INTERNAL_ERROR),
-				 errmsg("Could not drop schema created under PostgreSQL dialect: \"%s\"", schemaname)));
+		if (!missing_ok)
+			ereport(ERROR,
+					(errcode(ERRCODE_INTERNAL_ERROR),
+						errmsg("Could not drop schema created under PostgreSQL dialect: \"%s\"", schemaname)));
 		return;
 	}
